# openapi.yml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /send-money:
    post:
      summary: Send Money
      operationId: send_money_send_money_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMoneyRequestResponse"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMoneyRequestResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /accounts:
    get:
      summary: Get Account
      operationId: get_account_accounts_get
      security:
        - HTTPBearer: []
      parameters:
        - name: account_id
          in: query
          required: false
          schema:
            type: integer
            nullable: true
            title: Account Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
                title: Response Get Account Accounts Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /activities:
    get:
      summary: Get Activity
      operationId: get_activity_activities_get
      security:
        - HTTPBearer: []
      parameters:
        - name: activity_id
          in: query
          required: false
          schema:
            type: integer
            nullable: true
            title: Activity Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityResponse"
                title: Response Get Activity Activities Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /accounts-balance/{account_id}:
    get:
      summary: Get Account Balance
      operationId: get_account_balance_accounts_balance__account_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
            title: Account Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAccountBalanceResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /accounts/{account_id}:
    put:
      summary: Create Or Update Account
      operationId: create_or_update_account_accounts__account_id__put
      security:
        - HTTPBearer: []
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
            title: Account Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/AccountResponse"
                  - type: object
                title: Response Create Or Update Account Accounts  Account Id  Put
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /users:
    post:
      summary: Register User
      operationId: register_user_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterUserResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    get:
      summary: Get User
      operationId: get_user_users_get
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            nullable: true
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
                title: Response Get User Users Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /login:
    post:
      summary: Login
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserRequest"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginUserResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    AccountResponse:
      properties:
        account_id:
          type: integer
          title: Account Id
        baseline_balance:
          type: number
          title: Baseline Balance
        activity_window:
          items:
            $ref: "#/components/schemas/ActivityResponse"
          type: array
          title: Activity Window
      type: object
      required:
        - account_id
        - baseline_balance
        - activity_window
      title: AccountResponse
    ActivityResponse:
      properties:
        activity_id:
          type: integer
          nullable: true
          title: Activity Id
        owner_account_id:
          type: integer
          title: Owner Account Id
        source_account_id:
          type: integer
          title: Source Account Id
        target_account_id:
          type: integer
          title: Target Account Id
        timestamp:
          type: string
          title: Timestamp
        money:
          type: number
          title: Money
      type: object
      required:
        - activity_id
        - owner_account_id
        - source_account_id
        - target_account_id
        - timestamp
        - money
      title: ActivityResponse
    AddressRequestResponse:
      properties:
        street_name:
          type: string
          title: Street Name
        street_number:
          type: integer
          title: Street Number
        city:
          type: string
          title: City
        postal_code:
          type: string
          title: Postal Code
        country:
          type: string
          title: Country
      type: object
      required:
        - street_name
        - street_number
        - city
        - postal_code
        - country
      title: AddressRequestResponse
    GetAccountBalanceResponse:
      properties:
        account_id:
          type: integer
          title: Account Id
        balance:
          type: number
          title: Balance
      type: object
      required:
        - account_id
        - balance
      title: GetAccountBalanceResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginUserRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - password
      title: LoginUserRequest
    LoginUserResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
        - access_token
      title: LoginUserResponse
    RegisterUserRequest:
      properties:
        user_id:
          type: integer
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        address:
          $ref: "#/components/schemas/AddressRequestResponse"
      type: object
      required:
        - user_id
        - name
        - email
        - password
        - address
      title: RegisterUserRequest
    RegisterUserResponse:
      properties:
        user_id:
          type: integer
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        address:
          $ref: "#/components/schemas/AddressRequestResponse"
        status:
          type: string
          title: Status
      type: object
      required:
        - user_id
        - name
        - email
        - address
        - status
      title: RegisterUserResponse
    SendMoneyRequestResponse:
      properties:
        source_account_id:
          type: integer
          title: Source Account Id
        target_account_id:
          type: integer
          title: Target Account Id
        amount:
          type: number
          title: Amount
      type: object
      required:
        - source_account_id
        - target_account_id
        - amount
      title: SendMoneyRequestResponse
    UpdateAccountRequest:
      properties:
        baseline_balance:
          type: number
          title: Baseline Balance
      type: object
      required:
        - baseline_balance
      title: UpdateAccountRequest
    UserResponse:
      properties:
        user_id:
          type: integer
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        address:
          $ref: "#/components/schemas/AddressRequestResponse"
        status:
          type: string
          title: Status
      type: object
      required:
        - user_id
        - name
        - email
        - address
        - status
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer